+++ tests/ADN.Helpers.Tests/Models/Data/StringHelperTest.cs
@@ -1,5 +1,6 @@
 ﻿using ADN.Helpers.Data;
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -41,6 +42,27 @@ namespace ADN.Helpers.Tests.Data
         {
             var result = value.Right(length);
             Assert.Equal(expected, result);
+        }
+
+        [Theory]
+        [ClassData(typeof(TrimNonPrintableAsciiData))]
+        public void TrimNonPrintableAscii(string value, string expected)
+        {
+            var result = value.TrimNonPrintableAscii();
+            Assert.Equal(expected, result);
+        }
+
+        public class TrimNonPrintableAsciiData : IEnumerable<object[]>
+        {
+            public IEnumerator<object[]> GetEnumerator()
+            {
+                yield return new object[] { "", "" };
+                yield return new object[] { "abcdefghij", "abcdefghij" };
+                yield return new object[] { (char)0x12 + "a" + (char)0x142 + "b" + (char)0x242, "ab" };
+                yield return new object[] { " a~", " a~" };
+            }
+
+            IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
         }
     }
 }